move test to cxx2b-deducing-this
